Quality of life fixes.

+++ src/main/java/nl/devpieter/falsereality/Settings/Config.java
@@ -3,7 +3,7 @@ package nl.devpieter.falsereality.Settings;
 public class Config {
 
     public static boolean CustomTimeEnabled;
-    public static int CustomTime = 13000;
+    public static long CustomTime = 13000;
 
     public static boolean TimeSpedUpEnabled;
     public static long SpedUpBy = 15;
